feat: 일정 메모 원시값을 객체로 포장
