+++ backend/src/main/java/com/allog/dallog/schedule/domain/Schedule.java
@@ -1,8 +1,6 @@
 package com.allog.dallog.schedule.domain;
 
-import com.allog.dallog.schedule.exception.InvalidScheduleException;
 import java.time.LocalDateTime;
-import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -14,11 +12,9 @@ import javax.persistence.Table;
 @Table(name = "SCHEDULES")
 public class Schedule {
 
-    private static final int MAX_MEMO_LENGTH = 255;
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long Id;
+    private Long id;
 
     @Embedded
     private Title title;
@@ -26,28 +22,21 @@ public class Schedule {
     @Embedded
     private Period period;
 
-    @Column(nullable = false)
-    private String memo;
+    @Embedded
+    private Memo memo;
 
     protected Schedule() {
     }
 
     public Schedule(final String title, final LocalDateTime startDateTime,
         final LocalDateTime endDateTime, final String memo) {
-        validateMemoLength(memo);
         this.title = new Title(title);
         this.period = new Period(startDateTime, endDateTime);
-        this.memo = memo;
-    }
-
-    private void validateMemoLength(String memo) {
-        if (memo.length() > MAX_MEMO_LENGTH) {
-            throw new InvalidScheduleException("일정 메모의 길이는 255를 초과할 수 없습니다.");
-        }
+        this.memo = new Memo(memo);
     }
 
     public Long getId() {
-        return Id;
+        return id;
     }
 
     public String getTitle() {
@@ -63,6 +52,6 @@ public class Schedule {
     }
 
     public String getMemo() {
-        return memo;
+        return memo.getValue();
     }
 }
